Add GeoPointLoader component and convert batch size from Long to Int

+++ src/main/scala/com/stuart/geohash/domain/repositories/GeoHashRepository.scala
@@ -3,10 +3,7 @@ package com.stuart.geohash.domain.repositories
 import com.stuart.geohash.domain.models.geohash.{GeoHash, GeoHashMaxPrecision}
 
 trait GeoHashRepository[F[_]] {
-
   def create(geoHash: GeoHash): F[GeoHash]
-
   def findBy(geoHashMaxPrecision: GeoHashMaxPrecision): F[Option[GeoHash]]
-
-  def findAll(page: Long, size: Long): F[List[GeoHash]]
+  def findAll(page: Int, size: Int): F[List[GeoHash]]
 }
